+++ test/collector.js
@@ -100,6 +100,34 @@ metatests.test('Collector: timeout', async (test) => {
   }
 });
 
+metatests.test('Collector: default values', async (test) => {
+  const defaults = { key1: 1 };
+
+  const dc1 = collect(['key1'], { defaults, timeout: 50 });
+  const dc2 = collect(['key1'], { defaults: {}, timeout: 50 });
+  const dc3 = collect(['key1', 'key2'], { defaults, timeout: 50 });
+  dc3.set('key2', 1);
+
+  setTimeout(() => {
+    dc1.set('key1', 2);
+    dc2.set('key1', 2);
+    dc3.set('key1', 2);
+  }, 100);
+
+  const result1 = await dc1;
+  test.strictSame(result1, defaults);
+  const result3 = await dc3;
+  test.strictSame(result3, { ...defaults, key2: 1 });
+
+  try {
+    await dc2;
+    test.error(new Error('Should not be executed'));
+  } catch (error) {
+    test.strictSame(error.message, 'The operation was aborted due to timeout');
+  }
+  test.end();
+});
+
 metatests.test('Collector: fail', async (test) => {
   const dc = collect(['key1']);